+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.87.1
 
+- Allow CallLinkRootKey to vary in size and remove call link epochs from backup