+++ RELEASE_NOTES.md
@@ -1,3 +1,2 @@
-v0.87.1
+v0.87.2
 
-- Make ROOT_KEY_MAX_BYTES_FOR_SHO visible to the outside world